DDCYLS - 7708 - upgrading stub repo to scala 3 (#71)
* upgrading stub repo to scala 3

* fixing it test failure and updating app dependencies

* adding scala fmt scala 3 dialect

* updating play framework

+++ app/uk/gov/hmrc/economiccrimelevystubs/services/EclRegistrationReferenceService.scala
@@ -38,7 +38,8 @@ class EclRegistrationReferenceService @Inject() (sequenceRepository: SequenceRep
 
   def getPreviousReferences: Future[List[String]] =
     sequenceRepository.getCurrentReference(eclRegistrationReferenceKey).map {
-      case Some(ref) => List.range(1, ref + 1).map(i => s"$prefix${pattern.format(i)}")
+      case Some(ref) =>
+        List.range(1, ref.toInt + 1).map(i => s"$prefix${pattern.format(i)}")
       case _         => List.empty
     }
 }
